added 10x10 nonogram
This has a 10x10 nonogram. I updated wisdom of crowds. There was a problem in the loop for choosing 0 and 1. I also updated single point crossover to do it at 3 points. mutation rate is now higher and mutates the number of blocks equal to half the board size. This is done in hope to cause diversity in the population.